+++ history.md
@@ -3,6 +3,8 @@
 * Fixed issues where notNode was not retracting all contexts when a fact was retracted
 * Fixed issue [50](https://github.com/C2FO/nools/issues/50) in compiling rule with or condition
 * Fixed issue [53](https://github.com/C2FO/nools/issues/53) in constraints with hash refrences.
+* Merged pull request [49](https://github.com/C2FO/nools/pull/49)
+* Added tests for logicalNot and truthy statements
 
 # 0.1.8 / 2013-08-14